+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - TELEKINESIS description now includes throw distance. [#1450](https://github.com/dmdorman/hero6e-foundryvtt/issues/1450)
 - Fixed issue with negative RECOVERY during "Take A Recovery". [#1396](https://github.com/dmdorman/hero6e-foundryvtt/issues/1396)
 - 5e HARDENED now works against PENETRATING. [#1437](https://github.com/dmdorman/hero6e-foundryvtt/issues/1437)
+- Improved token selection when there are 2+ tokens associated with the same actor. [#1397](https://github.com/dmdorman/hero6e-foundryvtt/issues/1397)
 
 ## Version 4.0.8